plaid_integration/requirements.txt
# Add your project dependencies here
plaid-python==8.4.0
Flask==2.3.3
python-dotenv==1.0.0
